+++ src/thread/Timer.hh
@@ -3,18 +3,20 @@
 #ifndef TIMER_HH
 #define TIMER_HH
 
+#include <cstdint>
+
 namespace openmsx {
 
 namespace Timer {
 
 	/** Get current (real) time in us. Absolute value has no meaning.
 	  */
-	unsigned long long getTime();
+	uint64_t getTime();
 
 	/** Sleep for the specified amount of time (in us). It is possible
 	  * that this method sleeps longer or shorter than the requested time.
 	  */
-	void sleep(unsigned long long us);
+	void sleep(uint64_t us);
 
 } // namespace Timer
